styles fixes
